telegramrequirements.txt
python-telegram-bot
Pillow
ImageHash
python-dotenv